elf_core_file_p(): also load all elf section headers
The minicoredumper adds a section note to the core file with
information about which parts of the core were dumped. This
information should be made available to bfd_core.

This is mostly copy/paste from elf_object_p() to add support for
loading all section headers. Now that the functions are so
similar, it make sense to unify them.

Signed-off-by: John Ogness <john.ogness@linutronix.de>